update - createMapDialog + resourceView

+++ src/main/java/lu/innocence/ignis/component/ChessBGDrawer.java
@@ -8,7 +8,7 @@ import javafx.scene.paint.Color;
  */
 public class ChessBGDrawer {
 
-    public static void drawChessBackground(GraphicsContext g, double width, double height, int cellWidth, int cellHeight) {
+    public static void drawChessBackground(GraphicsContext g, int width, int height, int cellWidth, int cellHeight) {
         for (int i = 0; i < width; i++) {
             for (int j = 0; j < height; j++) {
                 drawChessBackgroundSingle(g, i, j, cellWidth, cellHeight);
